refactor(elementor): update namespaces and apply Singleton trait
